Added prime function to helpers
