+++ test/helpers.test.js
@@ -29,4 +29,22 @@ describe('#helpers', () => {
             assert(!lister.odd(Math.PI))
         })
     })
+
+    describe('#prime', () => {
+        it('should return true when given zero', () => {
+            assert(lister.prime(0))
+        })
+
+        it('should return true when given one', () => {
+            assert(lister.prime(1))
+        })
+
+        it('should return false when given 2', () => {
+            assert(lister.prime(2))
+        })
+
+        it('should return true when given a prime > 2', () => {
+            assert(lister.prime(41))
+        })
+    })
 })
